move instrumentation to JSBackendDiffMaker
