+++ hkmc2/shared/src/test/mlscript/staging/Syntax.mls
@@ -11,10 +11,11 @@ staged module A
 
 :js
 :slot
+:ds
 :w
 staged module A
 //│ ╔══[WARNING] `staged` keyword doesn't do anything currently.
-//│ ║  l.15: 	staged module A
+//│ ║  l.16: 	staged module A
 //│ ╙──      	       ^^^^^^^^
 //│ Pretty Lowered:
 //│   define staged class A in set block$res = undefined in end
