+++ hkmc2DiffTests/src/test/scala/hkmc2/JSBackendDiffMaker.scala
@@ -107,7 +107,10 @@ abstract class JSBackendDiffMaker extends MLsDiffMaker:
           new JSBuilder
             with JSBuilderArgNumSanityChecks
       val resSym = new TempSymbol(S(blk), "block$res")
-      val lowered0 = low.program(blk)
+      val lowered0 = if stageCode.isSet then
+        val instrumentation = new Instrumentation
+        instrumentation.transform(low.program(blk))
+        else low.program(blk)
       val le = lowered0.copy(main = lowered0.main.mapTail:
         case e: End =>
           Assign(resSym, Value.Lit(syntax.Tree.UnitLit(false)), e)
